Implement separate state reads observation for layout modifiers
Previously the state reads done within the layout blocks of LayoutModifiers on the child layout nodes were observed as part of the layoutBlock invocation of the parent's MeasurePolicy. It is not only causes us to do extra work as we rerun the parent block when we can only rerun the child's block, but also it is not entirely correct as the child node can be individually scheduled for relayout. Which means MeasureAndLayoutDelegate will only call layoutNode.replace() without rerunning the whole parents layout block meaning that we were doing reinvocation of the LayoutModifiers block without any state reads observations.
In order to fix that first I tried to move the whole layout related state reads observation from LayoutNode.layoutChildren() into OuterMeasurablePlaceable.placeAt() similarly to how it is done for the observation of the measure block (in OuterMeasurablePlaceable.remeasure()). However it is not solving all the use cases as LayoutNode.layoutChildren() could be run not as part of the OuterMeasurablePlaceable.placeAt() flow but during the calculateAlignmentLines() call. I didn't manage to figure how to reimplement this flow differently. Instead I decided to introduce extra observation scope for LayoutModifier's layout blocks. So now we track the state reads of modifier`s blocks and the node's block individually, but as a result of the state change we do the same layoutNode.requestRelayout(). This also gives us possibility to not re-run parent's layout block in such cases.

Test: new tests in AndroidLayoutDrawTest and ModelReadsTest
Bug: 192595563
Change-Id: I83c0cc5ab89dfe0eee85fda1b1fb7297a3129ce4