Ensure initial destinations reach RESUMED when deeplinking
Rather than using the backStack size as the indicator
on whether we're doing the initial composition
of the Crossfade, remember a boolean variable to track
this state. This ensures that even when deep linking
to a destination on the initial composition (i.e.,
going from 0 to 2+ destinations) that we still
mark the transition as complete.

Relnote: "Fixed an issue in Navigation Compose where
a deep linked `NavBackStackEntry` would not reach the
`RESUMED` state."
Test: new NavHostTest test passes

Change-Id: I192c5d5922901f94ea0cf5abe7a922c0a63cc033