Added support for SwipeToDismiss to Wear Compose.
SwipeableState and the swipeable modifier are copies of their equivalents
under Compose/Material, with two additions:
1) the swipe offset can be reset to its initial value after swiping
2) optional callback when animations have run and the offset has been reset.
Integration with Androidx navigation will follow separately.
Added demo of swipe-to-dismiss to the Wear Compose integration demos.

Test: androidx.wear.compose.integration.demos.test - more detailed tests to follow.

RelNote: "Added a SwipeToDismissBox and swipeable modifier that can be used to
support a left-to-right swipe-to-dismiss gesture.
Although independent of navigation components,
this is expected to be used to exit one screen and navigate to another. Added
integration tests to demo swipe-to-dismiss."

Bug: 190807006

Change-Id: I7bbaa6256c170c7e46cd9d7a8de5af0ea45d2a06