Desktop. Introduce common WindowScope for Window and Dialog
The old WindowScope renamed to FrameWindowScope, DialogScope to DialogWindowScope.
Usual Window and Dialog are both windows, but with different decoration and behaviour.

Introducing a common WindowScope allows to write a common logic for Window/Dialog.
If users want individual logic, they can use FrameWindowScope or DialogWindowScope.

Example of using WindowScope you can find in the
https://android-review.googlesource.com/c/platform/frameworks/support/+/1773990
(WindowDraggableArea)

Test: manual ./gradlew :compose:desktop:desktop:desktop-samples:run1 -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I3dcc07151a735d50b966957e7f04b3794376f162