Add UserStyle get by id
UserStyle contains a Map which may be mutable this means we can't
easily maintain a parallel mapping of Id to Option and therefour
the new get by id has to be O(n). We expect style maps to be small
so this shouldn't be a performance issue.

Test: Added a new test
Bug: 194464508
Relnote: We have added an overloaded UserStyle.get which accepts UserStyleSetting.Id.
Change-Id: I2aa0ff7c6075a91436fe9b5a87f935a1356862e2