Disallow duplicated style options ids.
The option ids should be unique within a given setting, as they are
used to uniquely identify the option in a few places. This changelist
enforces this with a require in the constructor of the
UserStyleSetting subclasses that need it.

Bug: 194486716
Test: Unit tests
Relnote: 'We now reject any user style schema that has settings or options with conflicting IDs'
Change-Id: Ic27153688978dca2bd801b4d69af6296a2eb3b2c