Pin Compose Runtime Dependencies to Compose 1.0.0
This is the first part in decoupling the Compose groups
so they can release independently.  The Runtime and Compiler
must remain tip-of-tree project dependencies because 1.0.0
is not compatible with Kotlin 1.5.21, so they were
not updated.  They can be updated to 1.0.1 in the future.

Additionally, the intra-group dependencies were kept as
project dependencies.

Updates to the latest/tip-of-tree versions should only
happen when needed.

Test: gw bOS
Test: gw compileReleaseKotlin && gw compileReleaseKotlin -Pandroidx.compose.multiplatformEnabled=true
Change-Id: I4da89b14446606cf9477b31c00e3d7d8b215c68c