Ensure the NavigatorState.backStack is set before onAttach
When restoring the state in onGraphCreated(), we should
ensure that all of the NavigatorState is already
populated before we call onAttach() on the
Navigator and provide it the populated NavigatorState.

This ensures that Navigators that read the state
directly in onAttach (such as DialogFragmentNavigator)
get the correct state immediately and can use that
to reconnect to the DialogFragments.

Relnote: "Fixed an issue where dismissing a
`DialogFragment` would not correctly update the
`NavController` state after recreating your activity
(i.e., after a configuration change)."
Test: new NavHostFragmentTest passes
BUG: 195423173

Change-Id: Icd72b58ece54484b576b44fc2fb9dbf68ccb9ab9