Add support for glowing overscroll.
This CL adds support for pre-stretch glowing overscroll in all types of default scrolling containers (LazyColumn, LazyRow, Modifier.verticalScroll, Modifier.horizontalScroll).

With the addition of the effect itself, this CL adds experimental android-only API to allow for customisations of this effect, for functionality to be on par with views glowing effect.

Bug: 171682480
Test: add new intergration and screenshot tests
Relnote: Glow effect for scroll has been added. New experimental OverScrollConfiguration API has been added to allow for configuration of the overscroll visual effect. Provide null to turn off the overscroll effect.
Change-Id: I0c3042f89365408e92f657330ed3164e6f3e12f8