Move ProfileInstaller to write a skip file.
This replaces the previous heuristics behavior that attempted to decide
if a profile should be written, which turned out to be unreliable on
some API levels.

The new behavior is:
 - ProfileInstaller will write the PackageInfo.lastUpdatedTime to a file
 in the apps files dir {GUID}.profileinstaller on success
 - Before read/transcode, the file is checked to see if it contains the
 same lastUpdatedTime as the last profile to be written.

This means that a profile will only be installed once per lastUpdatedTime
in PackageInfo, which corresponds with app installs or upgrades.

Bug: b/195603861
Test: ProfileInstallerTest
Relnote: "Change profileinstaller skip behavior to log the
PackageInfo.lastUpdatedTime in a file in the app's files directory and
compare it prior to installing the profile on the next run."

Change-Id: Ib93d1917a8a6a56162bf9e84c58ecc585fe86bfb