Fix issue with label being transparent if color provided in typography
`lerp(Color.Unspecified, color, 0)` returns `Color(0, 0, 0, 0, validColorSpace)`. Unlike Color.Unspecified, that is treated as an absent color, that one will be a transparent color.
In text field we lerp between `subtitle1` and `caption` for Label. If color is provided to `caption` typography, the result from lerp will be transparent color. And Text will treat it as a valid color and not take color from the LocalContentColor. So instead we check such a case and fallback to content color before lerping if needed.

Test: new tests
Fixes: 194759511
Change-Id: Ib8b1f4c39c4238bb75a64878799cd6705b1fd300