Consider Lifecycle when composing entries in NavHost
It is possible to get into a scenario where when the NavHost is first
composed, none of the entries have reached the STARTED LifecycleState.
This will cause us to never compose any content to the screen.

Instead of expecting entries to already be visible, we should observe
their lifecycles on every composition, and in response to lifecycle
changes we should trigger a recompose. This will ensure that whenever an
entry moves the the STARTED state we will have the opportunity to
compose its destination's content to the screen.

RelNote: "`NavHost` now observes for changes in the `Lifecycle.State` of
`NavBackStackEntry`s which means using a `NavHost` in a fragment no
properly be composed instead of resulting in a blank screen."
Test: existing tests pass and tested in samples
Bug: 195864489

Change-Id: I4eb8586e4db4bebfac63525e9dcdab11b57fcff5