Multi-process lock for opening DB (and migrations too)
Adds a lock at the FrameworkSQLite* level, to protect multi-process 1st time database creation and migrations. To avoid a performance regression, process-lock is only done when the database instance is not known to be opened, such that consecutive getDatabase() calls are quick and do not have to use a FileLock. Note that in-memory databases (null name) also don't use a FileLock.

Bug: 193182592
Test: n/a
Relnote: Added API for multi-process lock and usage at the FrameworkSQLite* level, to protect multi-process 1st time database creation and migrations.
Change-Id: Ied267cd32e3248cec5ebb772778d2e94fd450270