Implement Row, Column and BackgroundModifier for Wear.
This implements same-axis alignemnt for Rows and Columns (e.g.
horizontal alignemnt for a Row) by wrapping the Row in a Column to set
the alignment (and vice versa).

This also copies the findModifier() utility to the main package; it's
very useful in the composition translators too.

Test: included.
Relnote: n/a
Change-Id: Ia35296877e4b0c9c29d45ec3a5d1508ff0c11b8f