ScalingLazyColumnStateAdapter: fix up the calculation of decimalLastItemIndex() and decimalFirstItemIndex
There was a int/float conversion issue with ScalingLazyColumnStateAdapter.decimalLastItemIndex()/.decimalFirstItemIndex() that was resulting in the scroll indication for ScalingLazyColumns being more "lumpy" than it needed to be.

Also added some extra tests to PositionIndicatorTest to test for reverseLayout (it was these new tests that exposed the issue). Note there is also an underlying issue in LazyList (b/198751807) resulting in having to ignore one of the tests for small lists with vertical alignment reversed.

Test: androidx.wear.compose.material.PositionIndicatorTest
Bug: 198752976
Change-Id: I61cdfa502e8d4677ba38a8d567d6d3c94fc02274