Add ability to disable built-in converters
This CL adds a new property to the TypeConverters annotation to disable
built in converters in certain scopes. This can be useful when the
developer do not want to use built in converters and rather prefer a
compile time error.

I've also updated XProcessing message assertions to receive a list such
that it can handle cases where multiple messages match the given
message.

Bug: 73132006
Test: BuiltInConverterFlagsTest
Relnote: "We have added a new property to the TypeConverters annotation
to let developers disable built in enum and uuid converters. By default,
these converters are on but you can disable them for a certain scope, or
for the whole database. See TypeConverters documentation for details."

Change-Id: I671e8b1a8eb71d3309da04feaf6f9b18719489fc