setGraph() will now also clear saved back stacks
When you call setGraph() a second time with a
different graph, the back stack was popped. Now,
that code also clears all saved back stacks, ensuring
that the state is reset back to where it was as
if the first graph was not ever set.

Test: new NavControllerTest tests pass
BUG: 189979430
Relnote: "Calling `setGraph()` with a new graph
will now also clear any saved back stacks in addition
to its previous behavior of popping the back stack."

Change-Id: I97f96a65efea43d4341c9572c3c1c63f343ab6ed