Use {argName} for validating deep link arguments
While in many cases query parameters are in the
form of myarg={myarg}, this should never be a hard
requirement. Instead, only the value within braces
(what is actually used as the argument name) should
be used when validating that all required
arguments are present in the NavDeepLink.

Relnote: "Fixed a regression introduced in
[Navigation `2.4.0-alpha09`](#2.4.0-alpha09) when using
deep links with query parameters where Navigation would
incorrectly validate argument names using the query
parameter name (i.e., the `id` of `?id={userId}`) rather
than using the actual argument names present in the value
(the `userId` in the example above)."
Test: new NavDestinationAndroidTest tests pass
BUG: 200845660

Change-Id: Id2f399a5e19e593f34e02fc811f1cc28ec0689b8