Remove currentWindowMetrics api.
Remove current window metrics API since we can not provide the value
quickly and consistently. We saw that usinc
Activity#registerComponentCallback registered for component callbacks of
the Application. We also tried to register a layout change listener but
that was too slow. We considered adding a view to pipe configuration
changes but that is not guaranteed to work.
Use WindowMetricsCalculator instead.

Relnote: Remove currentWindowMetrics use WindowMetricsCalculator instead
Bug: 201101222
Test: n/a
Change-Id: Icda5f3802692e70d0220a491c90999866a1b73cb