Change time between injected input events to 16ms
This is more realistic and aligns better with the rest of the test
infrastructure that assumes frames are 16ms.

Fix: 195536192
Test: ./gradlew bOS
Relnote: "The default time between injected input events has been
changed from 10ms to 16ms. This potentially changes the outcome of tests
that perform input gestures, like a specific swipe."

Change-Id: I829fd33dd4ba14313ac73daf7c2e770955c90df6