Make watchface animation smoother
The previous computeDelayTillNextFrame plus choreographer frames
was not as regular as it could have been, because with
high framerates, computeDelayTillNextFrame's phase computation was
fighting with the choreographer.  Its still useful for very low
rendering rates (e.g. 1hz where rendering should be aligned to the
top of the second).

Where possible we don't post a delayed task, relying on the
choreographer for the delay.  Past a threshold we use a delay but
aim to run the task to post a choreographer frame well in advance
to ensure there's enough time for it to be scheduled.

In addition we add tests for ambient -> interactive and
interactive -> ambient.

Test: Mnual testing with perfetto
Change-Id: I9bbf2882a64ed7eebd711a51604cf230bac61237