Update samples code to make it easier to develop
Add a recycler view to house all the demo information and launch button.
This will make it easier to add new demos since it requires creating a
DemoItem instead of modifying a layout.
Convert DisplayFeaturesActivity to use a RecyclerView. This makes it
easier to process the logs since you can scroll through them.

Bug: 202325379
Test: manual
Change-Id: Ieefcb44a1525ff23caea09376af79644f53aed1f