Remove MIME type check in InputConnectionCompat.commitContent
For the OnReceiveContentListener API, the MIME types declared by the app
specify the types of content that are well-handled by the app; however
even if the content is of a different type, it is still passed to the
app to provide an opportunity for app-specific fallback logic (e.g. the
app could attempt to convert the content, or simply show a user-friendly
message indicating that the content is not supported).

For the IME commitContent API in the SDK, there is an anlogous contract:
the IME can try to call commitContent with content whose type is not
among the list of MIME types declared by the app. From the javadoc on
InputConnection.commitContent: "Editor can ignore this request when the
MIME type specified in inputContentInfo does not match any of
EditorInfo#contentMimeTypes."

In the SDK's implementation of the commitContent API, there are no
checks for the MIME type of the content. But in the AndroidX impl of the
API in InputConnectionCompat.commitContent, there is logic to filter out
content whose type doesn't match any of the declared MIME types. This
makes it inconsistent with the platform behavior of commitContent and
with the OnReceiveContentListener API.

In this PR, the logic in InputConnectionCompat.commitContent is updated
to no longer filter on the MIME type and let the receiving app decide
how to handle the content.

Bug: 193820646
Test: InputConnectionCompatTest
Test: AppCompatEditTextReceiveContentTest
Change-Id: I5bdd99a34b67832b2fcc9cdb7e70f2925912830e